+++ block-streamer/src/graphql/queries/get_bitmaps_exact.rs
