+++ block-streamer/src/graphql/queries/get_bitmaps_wildcard.rs
