+++ block-streamer/src/main.rs
@@ -3,7 +3,6 @@ use tracing_subscriber::prelude::*;
 mod block_stream;
 mod delta_lake_client;
 mod graphql;
-mod graphql_queries;
 mod indexer_config;
 mod lake_s3_client;
 mod metrics;
